Validação para não permitir deletar produto se ele estiver em um pedido e ajuste de layout nos componentes do pedido

+++ back/src/main/java/com/bruno/pedidos/controller/ProdutoController.java
@@ -97,6 +97,9 @@ public class ProdutoController {
         HttpResponse response = new HttpResponse();
 
         try {
+            if(produtoRepository.verificaPedidoProduto(id) > 0) {
+                throw new Exception("Erro ao deletar o Produto. Motivo: Ele possui pedidos cadastrados no sistema.");
+            }
             produtoRepository.deleteById(id);
             response.setStatus(HttpStatus.OK);
             response.setMessage("Produto deletado com sucesso.");
